p3

+++ src/tests/filesys/seq-test.c
@@ -29,7 +29,7 @@ seq_test (const char *file_name, void *buf, size_t size, size_t initial_size,
 
       ofs += block_size;
       if (check_func != NULL)
-        check_func (fd, ofs);
+        check_func (fd, (long) ofs);
     }
   msg ("close \"%s\"", file_name);
   close (fd);
