+++ src/tests/vm/pt-grow-stk-sc.c
@@ -14,19 +14,19 @@ void
 test_main (void)
 {
   int handle;
-  int slen = strlen (sample);
+  int slen = (int) (strlen (sample));
   char buf2[65536];
 
   /* Write file via write(). */
-  CHECK (create ("sample.txt", slen), "create \"sample.txt\"");
+  CHECK (create ("sample.txt", (unsigned int) slen), "create \"sample.txt\"");
   CHECK ((handle = open ("sample.txt")) > 1, "open \"sample.txt\"");
-  CHECK (write (handle, sample, slen) == slen, "write \"sample.txt\"");
+  CHECK (write (handle, sample, (unsigned int) slen) == slen, "write \"sample.txt\"");
   close (handle);
 
   /* Read back via read(). */
   CHECK ((handle = open ("sample.txt")) > 1, "2nd open \"sample.txt\"");
-  CHECK (read (handle, buf2 + 32768, slen) == slen, "read \"sample.txt\"");
+  CHECK (read (handle, buf2 + 32768, (unsigned int) slen) == slen, "read \"sample.txt\"");
 
-  CHECK (!memcmp (sample, buf2 + 32768, slen), "compare written data against read data");
+  CHECK (!memcmp (sample, buf2 + 32768, (unsigned int) slen), "compare written data against read data");
   close (handle);
 }
